+++ src/components/slide/content/poster/poster-subscription.jsx
@@ -21,12 +21,12 @@ function PosterSubscription({
 }) {
   const { t } = useTranslation("common", { keyPrefix: "poster-selector-v2" });
 
-  const [subscriptionEvents, setSubscriptionEvents] = useState(null);
+  const [subscriptionOccurrences, setSubscriptionOccurrences] = useState(null);
   const [loadingResults, setLoadingResults] = useState(false);
 
   const [firstAdmin] = admin;
   const optionsEndpoint = firstAdmin.endpointOption ?? null;
-  const searchEndpoint = firstAdmin.endpointSearch ?? null;
+  const subscriptionEndpoint = firstAdmin.endpointSubscription ?? null;
 
   const {
     subscriptionNumberValue = [],
@@ -45,8 +45,7 @@ function PosterSubscription({
 
   const subscriptionFetch = () => {
     const query = new URLSearchParams({
-      type: "events",
-      itemsPerPage: subscriptionNumberValue,
+      numberOfItems: subscriptionNumberValue,
     });
 
     const places = subscriptionPlaceValue.map(({ value }) => value);
@@ -69,12 +68,12 @@ function PosterSubscription({
 
     setLoadingResults(true);
 
-    fetch(`${searchEndpoint}?${query}`, {
+    fetch(`${subscriptionEndpoint}?${query}`, {
       headers: getHeaders(),
     })
       .then((response) => response.json())
       .then((data) => {
-        setSubscriptionEvents(data);
+        setSubscriptionOccurrences(data);
       })
       .finally(() => {
         setLoadingResults(false);
@@ -121,23 +120,23 @@ function PosterSubscription({
                   </tr>
                 </thead>
                 <tbody>
-                  {subscriptionEvents?.length > 0 &&
-                    subscriptionEvents?.map(
+                  {subscriptionOccurrences?.length > 0 &&
+                    subscriptionOccurrences?.map(
                       ({
-                        entityId,
-                        occurrences,
-                        imageUrls,
+                        eventId,
+                        imageThumbnail,
+                        image,
+                        startDate,
+                        endDate,
                         title,
                         organizer,
+                        place,
                       }) => {
-                        const firstOccurrence =
-                          occurrences.length > 0 ? occurrences[0] : null;
-
                         return (
-                          <tr key={`event-${entityId}`}>
+                          <tr key={`event-${eventId}`}>
                             <td>
                               <img
-                                src={imageUrls?.small}
+                                src={imageThumbnail ?? image}
                                 alt={title}
                                 style={{ maxWidth: "80px" }}
                               />
@@ -147,18 +146,11 @@ function PosterSubscription({
                               <br />
                               {organizer?.name}
                             </td>
+                            <td>{place?.name}</td>
                             <td>
-                              {firstOccurrence && firstOccurrence.place?.name}
-                            </td>
-                            <td>
-                              {firstOccurrence && (
-                                <>
-                                  {`${formatDate(
-                                    firstOccurrence.start,
-                                    "L"
-                                  )} - ${formatDate(firstOccurrence.end, "L")}`}
-                                </>
-                              )}
+                              {formatDate(startDate, "L HH:mm")}
+                              {" - "}
+                              {formatDate(endDate, "L HH:mm")}
                             </td>
                           </tr>
                         );
@@ -194,6 +186,7 @@ PosterSubscription.propTypes = {
         endpointEntity: PropTypes.string,
         endpointOption: PropTypes.string,
         endpointSearch: PropTypes.string,
+        endpointSubscription: PropTypes.string,
       })
     ),
   }).isRequired,